requirements.txt
# LexGuard — Python Dependencies
# Install: pip install -r requirements.txt

streamlit>=1.30.0
google-genai>=1.0.0
python-dotenv>=1.0.0
snowflake-connector-python>=3.6.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
PyMuPDF>=1.23.0
tqdm>=4.65.0
pytest>=7.4.0
rank_bm25>=0.2.2
sentence-transformers>=2.2.2
